Crate google_healthcare1_beta1

This documentation was generated from Cloud Healthcare crate version 2.0.0+20210317, where 20210317 is the exact revision of the healthcare:v1beta1 schema built by the mako code generator v2.0.0.

Everything else about the Cloud Healthcare v1_beta1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

locations datasets annotation stores annotations create, locations datasets annotation stores annotations delete, locations datasets annotation stores annotations get, locations datasets annotation stores annotations list, locations datasets annotation stores annotations patch, locations datasets annotation stores create, locations datasets annotation stores delete, locations datasets annotation stores evaluate, locations datasets annotation stores export, locations datasets annotation stores get, locations datasets annotation stores get iam policy, locations datasets annotation stores import, locations datasets annotation stores list, locations datasets annotation stores patch, locations datasets annotation stores set iam policy, locations datasets annotation stores test iam permissions, locations datasets consent stores attribute definitions create, locations datasets consent stores attribute definitions delete, locations datasets consent stores attribute definitions get, locations datasets consent stores attribute definitions list, locations datasets consent stores attribute definitions patch, locations datasets consent stores check data access, locations datasets consent stores consent artifacts create, locations datasets consent stores consent artifacts delete, locations datasets consent stores consent artifacts get, locations datasets consent stores consent artifacts list, locations datasets consent stores consents activate, locations datasets consent stores consents create, locations datasets consent stores consents delete, locations datasets consent stores consents delete revision, locations datasets consent stores consents get, locations datasets consent stores consents list, locations datasets consent stores consents list revisions, locations datasets consent stores consents patch, locations datasets consent stores consents reject, locations datasets consent stores consents revoke, locations datasets consent stores create, locations datasets consent stores delete, locations datasets consent stores evaluate user consents, locations datasets consent stores get, locations datasets consent stores get iam policy, locations datasets consent stores list, locations datasets consent stores patch, locations datasets consent stores query accessible data, locations datasets consent stores set iam policy, locations datasets consent stores test iam permissions, locations datasets consent stores user data mappings archive, locations datasets consent stores user data mappings create, locations datasets consent stores user data mappings delete, locations datasets consent stores user data mappings get, locations datasets consent stores user data mappings list, locations datasets consent stores user data mappings patch, locations datasets create, locations datasets deidentify, locations datasets delete, locations datasets dicom stores create, locations datasets dicom stores deidentify, locations datasets dicom stores delete, locations datasets dicom stores export, locations datasets dicom stores get, locations datasets dicom stores get iam policy, locations datasets dicom stores import, locations datasets dicom stores list, locations datasets dicom stores patch, locations datasets dicom stores search for instances, locations datasets dicom stores search for series, locations datasets dicom stores search for studies, locations datasets dicom stores set iam policy, locations datasets dicom stores store instances, locations datasets dicom stores studies delete, locations datasets dicom stores studies retrieve metadata, locations datasets dicom stores studies retrieve study, locations datasets dicom stores studies search for instances, locations datasets dicom stores studies search for series, locations datasets dicom stores studies series delete, locations datasets dicom stores studies series instances delete, locations datasets dicom stores studies series instances frames retrieve frames, locations datasets dicom stores studies series instances frames retrieve rendered, locations datasets dicom stores studies series instances retrieve instance, locations datasets dicom stores studies series instances retrieve metadata, locations datasets dicom stores studies series instances retrieve rendered, locations datasets dicom stores studies series retrieve metadata, locations datasets dicom stores studies series retrieve series, locations datasets dicom stores studies series search for instances, locations datasets dicom stores studies store instances, locations datasets dicom stores test iam permissions, locations datasets fhir stores create, locations datasets fhir stores deidentify, locations datasets fhir stores delete, locations datasets fhir stores export, locations datasets fhir stores fhir concept map-search-translate, locations datasets fhir stores fhir concept map-translate, locations datasets fhir stores fhir observation-lastn, locations datasets fhir stores fhir patient-everything, locations datasets fhir stores fhir resource-purge, locations datasets fhir stores fhir resource-validate, locations datasets fhir stores fhir capabilities, locations datasets fhir stores fhir conditional delete, locations datasets fhir stores fhir conditional patch, locations datasets fhir stores fhir conditional update, locations datasets fhir stores fhir create, locations datasets fhir stores fhir delete, locations datasets fhir stores fhir execute bundle, locations datasets fhir stores fhir history, locations datasets fhir stores fhir patch, locations datasets fhir stores fhir read, locations datasets fhir stores fhir search, locations datasets fhir stores fhir search-type, locations datasets fhir stores fhir update, locations datasets fhir stores fhir vread, locations datasets fhir stores get, locations datasets fhir stores get iam policy, locations datasets fhir stores import, locations datasets fhir stores list, locations datasets fhir stores patch, locations datasets fhir stores set iam policy, locations datasets fhir stores test iam permissions, locations datasets get, locations datasets get iam policy, locations datasets hl7 v2 stores create, locations datasets hl7 v2 stores delete, locations datasets hl7 v2 stores export, locations datasets hl7 v2 stores get, locations datasets hl7 v2 stores get iam policy, locations datasets hl7 v2 stores import, locations datasets hl7 v2 stores list, locations datasets hl7 v2 stores messages batch get, locations datasets hl7 v2 stores messages create, locations datasets hl7 v2 stores messages delete, locations datasets hl7 v2 stores messages get, locations datasets hl7 v2 stores messages ingest, locations datasets hl7 v2 stores messages list, locations datasets hl7 v2 stores messages patch, locations datasets hl7 v2 stores patch, locations datasets hl7 v2 stores set iam policy, locations datasets hl7 v2 stores test iam permissions, locations datasets list, locations datasets operations cancel, locations datasets operations get, locations datasets operations list, locations datasets patch, locations datasets set iam policy, locations datasets test iam permissions, locations get, locations list and locations services nlp analyze entities

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically ...

let r = hub.projects().locations_datasets_dicom_stores_studies_series_instances_frames_retrieve_frames(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_series_instances_frames_retrieve_rendered(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_series_instances_retrieve_instance(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_series_instances_retrieve_metadata(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_series_instances_retrieve_rendered(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_series_retrieve_metadata(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_series_retrieve_series(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_series_search_for_instances(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_retrieve_metadata(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_retrieve_study(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_search_for_instances(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_search_for_series(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_studies_store_instances(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_search_for_instances(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_search_for_series(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_search_for_studies(...).doit().await
let r = hub.projects().locations_datasets_dicom_stores_store_instances(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir__concept_map_search_translate(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir__concept_map_translate(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir__observation_lastn(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir__patient_everything(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir__resource_validate(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_capabilities(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_conditional_patch(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_conditional_update(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_create(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_delete(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_execute_bundle(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_history(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_patch(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_read(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_search(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_search_type(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_update(...).doit().await
let r = hub.projects().locations_datasets_fhir_stores_fhir_vread(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-healthcare1_beta1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.14"
hyper-rustls = "^0.22"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^5.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_healthcare1_beta1 as healthcare1_beta1;
use healthcare1_beta1::{Result, Error};
use std::default::Default;
use oauth2;
use healthcare1_beta1::CloudHealthcare;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = CloudHealthcare::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().locations_datasets_fhir_stores_fhir__concept_map_search_translate("parent")
             .url("amet.")
             .target("takimata")
             .system("amet.")
             .source("duo")
             .concept_map_version("ipsum")
             .code("gubergren")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub use api::CloudHealthcare;

pub use client::Result;

pub use client::Error;

pub use client::Delegate;

Modules

api

client